pyproject.toml
[tool.poetry]
name = "map-gpt-embeddings"
version = "0.0.1"
description = "`map-gpt-embeddings` is a Singer library for Open AI, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["AJ Steers, Meltano"]
keywords = [
    "ELT",
    "KnowledgeBase",
]
license = "Apache 2.0"
packages = [
    { include = "map_gpt_embeddings" },
]

[tool.poetry.dependencies]
python = "<3.12,>=3.8.1"
singer-sdk = { version="^0.30.0" }
fs-s3fs = { version = "^1.1.1", optional = true }
beautifulsoup4 = "^4.12.1"
openai = "^0.27.4"
langchain = "^0.0.133"
azure-identity = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8 = "^5.0.4"
darglint = "^1.8.1"
black = "^23.1.0"
pyupgrade = "^3.3.1"
mypy = "^1.0.0"
isort = "^5.11.5"
singer-sdk = { version="^0.30.0", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "map_gpt_embeddings"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
map-gpt-embeddings = 'map_gpt_embeddings.mappers:GPTEmbeddingMapper.cli'
